feat: 초기 데이터 세팅 함수 분리
